Refactor: 커서 모듈화
