+++ src/shared/components/GlobalCursor.css
@@ -1,15 +1,14 @@
-.cursor {
-  --size: 40px;
-
+.global-cursor {
   position: fixed;
-  z-index: 10;
+  z-index: 1000;
   top: 0;
   left: 0;
+  --size: 40px;
   width: var(--size);
   height: var(--size);
   border-radius: 50%;
   border: 1px solid rgba(255, 255, 255, 0.3);
-  background: rgba(255, 255, 255, 0.3);
+  background: rgba(255, 255, 255, 0.05);
   margin: calc(var(--size) * -0.5) 0 0 calc(var(--size) * -0.5);
   transition:
     transform 0.15s ease-out,
@@ -24,7 +23,7 @@
   }
 }
 
-.cursor2 {
+.global-cursor2 {
   position: fixed;
   top: 0;
   left: 0;
@@ -35,6 +34,6 @@
   border-radius: 50%;
   margin: -2.5px 0 0 -2.5px;
   backdrop-filter: none;
-  z-index: 11;
+  z-index: 1001;
   pointer-events: none;
 }
